pyproject.toml
[project]
name = "dodo"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Illia Puzanov", email = "zoilorys@gmail.com"},
]
dependencies = [
    "fastapi[standard]>=0.115.4",
    "python-multipart>=0.0.17",
    "jinja2>=3.1.4",
    "pdfminer-six>=20240706",
    "langchain>=0.3.7",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.1.4",
    "tiktoken>=0.8.0",
    "qdrant-client>=1.12.1",
    "sqlmodel>=0.0.22",
    "celery>=5.4.0"
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "fastapi dev src/dodo/main.py"
celery = "celery -A src.celery.celery worker -l info"

[tool.pdm]
distribution = false
